api/src/requirements.txt
Django==1.9.7
django-bootstrap-static==3.3.6
django-crispy-forms==1.6.0
django-filter==0.13.0
django-formtools==1.0
django-reversion==2.0.5
djangorestframework==3.4.0
dominate==2.2.0
httplib2==0.9.2
itsdangerous==0.24
Markdown==2.6.6
MarkupSafe==0.23
psycopg2==2.6.2
requests==2.10.0
SQLAlchemy==1.0.13
visitor==0.1.3
Werkzeug==0.11.10
